refactor(network): update supported chain id by track
