+++ src/constants/envVariables.ts
@@ -5,4 +5,14 @@ const { TRACK, METADATA_URL_OVERRIDE_OPTIMISM_KOVAN, METADATA_URL_OVERRIDE_OPTIM
  * METADATA_URL_OVERRIDE_OPTIMISM: metadata url override for Optimism Mainnet
  */
 
-export { TRACK, METADATA_URL_OVERRIDE_OPTIMISM_KOVAN, METADATA_URL_OVERRIDE_OPTIMISM }
+export enum Track {
+    DEV1 = "DEV1",
+    DEV2 = "DEV2",
+    CANARY = "CANARY",
+    BETA = "BETA",
+    PRODUCTION = "PRODUCTION",
+}
+
+const TYPED_TRACK: Track | undefined = TRACK as Track
+
+export { TYPED_TRACK as TRACK, METADATA_URL_OVERRIDE_OPTIMISM_KOVAN, METADATA_URL_OVERRIDE_OPTIMISM }
